nv-repos/projet_cancer/usecase/cloudrun/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-storage
google-cloud-aiplatform
google-auth
requests
firebase